Finish get_analogies function and run a few tests
The tests mainly consisted on double-checking for the match
between the results obtained with the current version of
PyVOLCANS and the Matlab version of VOLCANS, considering sets
of Top 10/20 analogue volcanoes with scheme A (equal weight)
and for the following target volcanoes: Fuego, Kilauea, Okmok,
Cleveland, and Gede-Pangrango. All tests were successful.