Add some formatted output to be displayed on the terminal
Note that this is a preliminary solution. The main reason to
implement this formatted output was to add the multi-criteria
analogy value in the outputs because, given two or more analogues
with the same value of analogy, the ordering in Python is not
entirely clear so double-checking with other documents, including
Tierz et al. (2019), can be difficult.