Complete get_analogy_percentile functions
The 'get_analogy_percentile' function was written up and the
'get_many_analogy_percentiles' updated to make it output, not
only a dictionary of percentiles, but also a dictionary of per-
centages of 'better analogues' (see Figs. 6, 7) in Tierz et al.
(2019).

Additionally, a few (non-structured) tests were run to check that
the values of the percentiles (and, hence, the percentages of
'better analogues' corresponded with the values provided by VOLCANS,
for schemes A, B, C in Tierz et al. (2019) and the following volcanoes:
Kilauea, Fuego, Sinabung, Aluto, Okmok and Copahue.

All tests have been successful to the 1-percentile level. That is,
one case had a difference of 1% in better analogues because the per-
centiles are actually calculated in a slightly different way by
Matlab and Python (NB. I have created an issue in the Gitlab repo-
sitory to solve this matter).