Add a priori analogues as input in the main
The script is now able to handle the input of a priori analogue
volcanoes as "--apriori volcano1 volcano2 volcano3" and calls the
'get_many_analogy_percentiles' function to print the percentage of
volcanoes in the GVP database, which are better analogues than the
provided a priori analogues, according to the PyVOLCANS calculations
(NB. For the particular weighting scheme considered in the run).
I have also slightly modified the aforementioned function, basically
in terms of print formatting of these 'better analogues'.