Update files that contain analogy and data matrices
I modified "base.py" accordingly, to load the new matrices.
I checked that both "pyvolcans" executable and "pytest -vs"
are working after these updates. I will commit and push to Gitlab.

Note: the new matrices were computed when the GVP data for
Quetrupillan volcano (Chile) were updated for its used in
VOLCANS. This was done in the context of the PhD project of
Isla Simmons, University of Edinburgh.