Fix small issue when passed weight is equal to default value
In the previous version, if the user changed N weights, as
input arguments of PyVOLCANS, but left one or more unchanged
(i.e. the value entered was equal to the default, 0.2), the
programme would throw an error, even if the total sum of weights
provided by the user was 1 (hence, appropriate).

I have not managed to think of a good solution here. In my view,
if the user is entering less than the total number of weights as
argument, the programme should crash if that sum is not equal to
1. I think the problem resides in defining the 'passed' arguments
as those that are different from the default value (0.2). I think
that we should define 'passed' as 'written' in the command line
by the user, if this is possible. Many thanks!