Move functionality to open GVP website to pyvolcans.py
This commit is actually made of several changes:

1. In pyvolcans_func.py: Make get_analogies() function to return
(for now) 'top_idx', that is: the indices of the top analogue
volcanoes to the target volcano.

2. In pyvolcans_func.py: Create a open_gvp_website() function which
only takes the indices mentioned above as argument, and opens the
GVP website for the top analogue volcano to the target volcano.

3. In pyvolcans.py: Return the indices for the top analogues
from the call to the get_analogies() function, and then use the
returned variable to call the open_gvp_website() function.

The very quick (and informal) tests run so far seem to show that
the programme still runs adequately. Nevertheless, proper acceptance
criteria will be created when the merge request for this branch will
be submitted.