Make get_analogies() function to return top analogues and volcano_name
I have also modified the returned variables in the call to the function
in pyvolcans.py. The idea is then calling write_csv() from pyvolcans.py,
using the returned variables (I know this is quite 'brute-force' but I
will start by seeing if it works).

Then, I will try to make the call to write_csv() as optional flag of the
programme.