Fix output name (count of top analogues) of the csv file
In get_analogies() function, we add +1 to the variable 'count'
(set as argument, default=10), to derive the {count} top analogue
volcanoes because the target volcano (volcano_input) is included
in the matrices that are searched to obtain the Top {count} analogue
volcanoes. Hence, to derive (say) the Top 10, we are searching for
10+1 analogues, including the target volcano.

There are probably better ways to deal with this issue. Eventually,
we should include the target volcano in the outputs but separately:
for example, in the first/last row, separated from the list of top
analogues by a blank line.