Change input arguments for get_analogies() function
In the previous version, PyVOLCANS worked with the variable
'weighted_total_analogy_matrix' (shape: 1439,1439) to obtain
the total analogy values for the target volcano ('my_volcano'),
and from there, it derived the Top x(=count) analogue volcanoes.

In this version, the function has 'volcans_result' as one of the
input arguments. The column ['total_analogy'] from this Pandas df
already contains the values associated with the target volcano,
which are then used to derive the Top x analogue volcanoes as before.