Change input arguments for 'get_percentiles' functions
In the previous version of PyVOLCANS, both 'get_analogy_percentile()'
and 'get_many_analogy_percentiles()' functions used the variable
'weighted_analogy_matrix' to perform the calculations that led to
obtaining the percentage of volcanoes in the GVP database that are
'better analogues' to the target volcano, compared to each of the
'a priori' analogues given by the user.

In the current version, these calculations are performed using the
Pandas data frame: 'volcans_result'. The required changes to the
function calls in 'pyvolcans.py' have also been implemented.

Informal, manual testing of the results before and after the changes
in this commit, for 4 target volcanoes, and 4 a priori analogues per
each target volcano (16 volcanoes in total), has been carried out,
and shows that the outputs of PyVOLCANS remain unchanged after the
changes. Nevertheless, more formal testing may be needed.