Make the 'write_result()' function more versatile
In previous versions, the function 'write_result()' (previously
called 'write_csv()') could only be used to write the optional csv
file that PyVOLCANS generates when the optional flag '-w' (for
'--write_csv_file') is chosen.

Now, the aforementioned function is expected to be able to write
the main PyVOLCANS outputs either to the standard output or to the
.csv file, by indicating the variable 'channel'. In this way, the
function can be called inside 'get_analogies()' to print the top
analogue volcanoes into the terminal.

I am also planning to generate another function 'write_result_verbose()'
which can print all single-criterion analogies, only when the verbose
mode is activated (otherwise, the default behaviour of PyVOLCANS would
be to print the total analogy only). However, I am not certain about
whether it would be better to have this functionality as another
condition of the same function 'write_result()'.