Change columns order in the non-verbose result
I think it makes sense to have the 'string variables' of the
result (name, country) as the two first columns, and then the
numerical variables (Smithsonian ID, total +/- single-criterion
analogyies). This may make the automatic reading of the output
csv file easier for other programs (e.g. Matlab), wouldn't it?