Avoid writing the csv file in verbose mode by default
Inside the 'output_result()' function, the behaviour of the programme
in the previous version of the code, was the following:

Given that the csv file was written (if the optional flag had been
selected) before the optional flag for the verbose mode was checked,
the csv file was always written to include all single-criterion
analogy values. Then, the programme checked for the verbose flag and
'routed' a verbose or non-verbose (i.e. only total analogy reported)
output to stdout.

Now, the verbose-flag check occurs before, and it just assigns a
value to the variable 'my_columns', which is then used both to write
the csv file (if selected) and to output the PyVOLCANS result to the
stdout, in either a verbose or non-verbose format.