Move (and expand) formating of csv filename to pyvolcans.py
In the previous version of the code, the 'preparation' of the
filename to use in the optional csv file that contains the top
{count} analogues to the target volcano was done inside the function
'output_result()' in pyvolcans_func.py

Now, these changes (i.e. removing apostrophes, points or commas from
the target volcano's name) are done in pyvolcans.py, and only if the
optional flag to write the csv file has been set to true.

Moreover, the csv filename now also reflects the weighting scheme
selected by the user to run PyVOLCANS. Consequently, the csv filename
is now a unique identifier of the three main variables of PyVOLCANS:
(1) target volcano, (2) number of top analogues, (3) weighting scheme
used to compute these analogues.