Fix (preliminarily) issues with float precision in sum_of_weights
Depending on the choice of weighting scheme from the user, an unwanted
error message could be thrown due to precision problems with the float
representation of the weights given to each volcanological criterion.

I tried to fix this using the module 'decimal' (see commented code in
this commit) but it could not fix the issue and, at the same time,
produce the intended behaviour when weights were given as repeating-
decimal fractions: e.g. 1/3, 1/6, etc.

With the current solution, any sum of weights that deviates more than
1e-10 from 1 is considered as a mistake in the weighting scheme
provided. Anything within 1-1e-10 and 1+1e-10 is treated as a valid
weighting scheme, and the programme runs accordingly.