Solve issue related to duplicated criterion weights
In the previous version of the code, if the user accidentally
introduced two values for the weight of a given volcanological
criterion, PyVOLCANS just overwrote the first N occurrences of
the weight with the last value provided. Hence, the following
raised an error linked to sum_of_weights being different from 1:

`pyvolcans Colo -Ts 1 -Ts 0.1`

But the following gave no error and ran the code normally:

`pyvolcans Colo -Ts 0.1 -Ts 1`

In the current version, a PyvolcansError is thrown when any of the
two cases above occur. In short, I have added the "action='append'"
to the 'parser.add_argument()' calls for each criterion, and then
the programme checks for the length of the generated lists of weights
inside the '_frac_to_float()' function, throwing the error any time
this length is higher than 1, and keeping the original value for the
weight if length is equal to 1.