Warn user of the occurrence of too many 'perfect analogues'
A new function called 'check_for_perfect_analogues()' has been created.
This function takes the main PyVOLCANS result (Pandas dataframe with
top analogue volcanoes to the target volcano) and assesses whether all
these top analogues share the same value of total analogy.

If they do, the programme raises a PyvolcansError which is caught by a
try-except block in 'pyvolcans.py'. This triggers a warning message but the
excecution of the code is not terminated. In other words, the top analogues
are still provided, then the user can take into consideration the relevance
of the results obtained, after having been warned of possible data deficiencies
and/or a weighting scheme 'too simplified' (e.g. search for a volcano on a
subduction zone under continental crust: there are hundreds of volcanoes exactly
like that one).