Fix open-website procedure to always open top-analogue website
In the previous version of the code, the variable 'top_analogues_idx',
derived from 'get_analogies()', was used to obtain the VNUM (or
Smithsonian ID) for the top analogue volcano. The VNUM was in turn
used to generate the appropriate URL link to the GVP website of the
top analogue volcano.
However, the target volcano had not been 'filtered out' from the
'top_analogues_idx' variable, and this generated some issues with
which website was actually opened when calling 'open_gvp_website()'.

In the current version, the variable 'top_analogues', which basically
represents the Pandas dataframe that contains all information for the
top analogue volcanoes, is used to get the desired VNUM instead. This
avoids the aforementioned issues and, in addition, simplifies the code.

We are now going to double-check whether the variable 'top_analogues_idx'
needs to be returned by 'get_analogies()' or not.