Remove 'top_idx' from the returned variables by 'get_analogies()'
The variable is not required outside the function as the variable
'filtered_result'/'top_analogues' contains all the information that
the programme needs to perform all its functionalities.

In previous versions, 'top_idx' was used to carry out the procedure
that opened the GVP website for the top analogue volcano. However,
this has now been improved and simplified, making the aforementioned
variable redundant outside 'get_analogies()'.