Fix 'a priori' procedure so target volcano can be a VNUM
In the previous version of the code, running PyVOLCANS using the
VNUM (or Smithsonian ID) to select the target volcano threw an error.
It was just a print error (str or int formats) and this has been
fixed.

Perhaps the only issue that remains as concerns the 'a priori'
procedure is that the programme is unable, at the moment, to deal
with a list of a priori analogues that is composed of a mix of
volcano names and VNUMs (or Smithsonian IDs). It may be unusual that
the user is particularly interested in running PyVOLCANS using such
mix of formats for the selected a priori analogues, but still, it
would be good to add this functionality to the programme.