Change message for test test_set_weights_from_args_error()
Given that we had modified the number of decimal places that the
function 'set_weights_from_args()' outputs in its error message,
the message we were checking against the error message was not
the same exactly, as so the test was failing. I just modified the
message to check and the test currently passes in my machine.