Modify np.isclose() call to produce more expected results
In the previous version of the code, we were not setting the relative
tolerance (rtol) of np.isclose() and that might have changed the be-
haviour we were expecting. This is because the actual formula that
np.isclose() is using is the following:

absolute(`a` - `b`) <= (`atol` + `rtol` * absolute(`b`))

where `a` is `sum_of_weights` and `b` is 1, in our case. By removing
the `rtol` term from the equation (i.e. setting rtol = 0), the True/False
output of the function call only depends on the absolute value of the
difference between `sum_of_weights` and 1. Otherwise, we had to account
for a 10e-5 factor (i.e. the default value for `rtol`) to be added to
`atol` in the formula above.