Add function to plot analogies for a priori analogues
A new function called `plot_bar_apriori_analogues()` has been created.
The function takes the list of a priori analogue volcanoes and the
VOLCANS results, and generates a bar plot that shows the values of
single-criterion and total analogy between the target volcano and any
a priori analogue volcano that the user has indicated.

At the moment, the function gets called in pyvolcans.py every time that
the programme is run. It is also called in a way that a .png file, with
the figure content in it, is automatically saved in each run.

The next steps are: (1) make the call of the function optional,
(2) make the saving of the figure optional (both via command-line
flags), (3) create another figure that plots the percentages of 'better'
analogues (to the target volcano) compared to the a priori analogues
provided by the user (please see Figs. 6, 7, in Tierz et al., 2019).