Skip any procedure related to a priori analogues if parameter is empty
In the previous version, when the user specified the optional flag
`--apriori`, but provided no a priori analogues, the programme would
perform the procedure to derive the percentage of 'better analogues',
including the introductory message to an empty list of a priori analogues
("According to PyVOLCANS, the following percentage of volcanoes in the
GVP database are better analogues to Unzendake than the 'a priori'
analogues reported below:").

With the addition of the plotting capability for these a priori
analogues, the programme crashed if `--apriori` was used together with
the optional flag `--plot_apriori`, but no a priori analogues were provided
(i.e. `my_apriori_volcanoes` was empty).

In the current version, the entire procedure related to the a priori
analogues (i.e. list of 'better analogues' + (optional) plotting)
is skipped when `my_apriori_volcanoes` is either None or empty.