Add function to format weights text and use it to write csv file
We have created a new function that takes the weights chosen by
the user to run PyVOLCANS, and returns a formatted text that can
be used (in the current commit) to write the filename of the csv
file where the top N analogues are stored (if the `-w` flag is
selected).

The same function will also serve to write the filenames for the
bar plots for 'a priori analogues' that are created by the function:
`plot_bar_apriori_analogues()`, in case the user decides to save
these plots by using the `--save_figures` flag.