Return plot handles from one plotting function
In this commit, two plot handles (one for the 'a priori analogues'
bar plot and another for the 'better analogues' bar plot) are
returned by the same 'plotting function': plot_bar_apriori_analogues().

In a second step, we will separate the two bar plots into two different
'plotting functions', each of which will return the corresponding
plot handle to then implement the testing.