Change the method to generate the pandas df from the dictionary
In the previous version, I believe there were some issues with the
way the dictionary of 'better analogues' was converted into a Pandas
dataframe. This led to errors in the plotting that is implemented in
the `plot_bar_better_analogues()` function, at least in my local
machine.

With the changes in this branch, the plotting seems to work well.
Some double-checking before merging will be required, nonetheless.