Fix issue with negative values for the criterion weights
In the previous version of the program, negative values for the
weights given to the volcanological criteria were accepted by
PyVOLCANS. As long as the sum of weights was 1, the program would
run as normal, even if the weighting scheme selected was fully
incompatible with the VOLCANS method and, hence, the set of analogue
volcanoes produced had no meaning. The issue has been fixed in this
commit.