Create a dictionary to check for criteria with data, for the target volcano
This dictionary contains each volcanological criterion as dict_key and
either 1 (there is data) or 0 (there is not data) as dict_value, all
considering the specific target volcano chosen by the user. The
dictionary will be a variable returned by the function:
calculate_weighted_analogy_matrix().

A quick check for volcanological criteria without any data, and the
output of appropriate warning messages for the user, could be done
in `pyvolcans.py` or in a separate function.