Move checking of missing data to pyvolcans_func.py
Instead of checking for volcanological criteria without available
data for the particular target volcano in pyvolcans.py, this is now
done inside the `calculate_weighted_analogy_matrix()` function inside
pyvolcans_func.py. In this way, we will not need to return the variable
`my_volcano_data_dictionary` in the aforementioned function (NB. To be
done in the next commit).