Print warning message only if weighting scheme uses non-data criteria
With this commit, the warning message related to volcanological criteria
for which there is no available data (for the specific target volcano)
is only printed if that particular criterion is used in the weighting
scheme selected by the user: that is, if the weight for that particular
volcanological criterion is greater than zero. Otherwise, no warning
message will appear because, strictly speaking, data are available
for all the selected criteria (those with weight > 0).