Load and preliminarily check volcano data
In this commit, the .mat files containing all the volcano-specific
data used to calculate all the single-criterion analogies for the
~1400 volcanoes in GVP v4.6.7 are loaded through `base.py`, and
subsequently, the volcano "ID profile" for the selected target
volcano is printed inside the `calculate_weighted_analogy_matrix()`
function (this is just a preliminary test of intended outputs. The
initial idea is to create a dedicated function that can print this
"ID profile" to the stdout, and maybe also produce a plot showing
these data in a more clear format, e.g. similar to Figure 3 in the
[VOLCANS
paper](https://link.springer.com/article/10.1007/s00445-019-1336-3/figures/3)).