Improve output formatting for the ID profile of target volcano
For now, the data are printed in the terminal, every time that
PyVOLCANS is run. The data are printed together with details on
the variables that are being printed, by volcanological criterion.
Float printed format is (maximum) 3 decimal places.

Until now, all the printing happens inside the function:
calculate_weighted_analogy_matrix(). We will move this to a separate
function in the following commits.