Add preliminary funcionality to output volcano data into csv
With this commit, the `output_volcano_data()` function is now
able to output a csv file that contains the 'ID profile' (or
volcanological data) of the volcano used to call the function.

The writing of the output csv file is conditional upon the selection
of an optional flag (to be set up in subsequent commits).