Remove deprecated option `interpolation` in np.percentile()
Function `get_analogy_percentile()` uses `np.percentile()` to calculate
the percentile that corresponds with the value of total analogy of a
given 'a priori' volcano entered by the user, within the vector of ALL
total-analogy values calculated for the target volcano against any other
volcano in the GVP database.

The option `interpolation` has been deprecated and running the PyVOLCANS
tests raised a Warning message. Also, removing the 'midpoint' option
implies that the default 'linear' is used instead, which approaches
better the method used in the original Matlab scripts of VOLCANS.